INC-224 refactor: rename usersService and types files
