+++ src/services/usersFollowingAndFollowersService/index.ts
@@ -1,0 +1,2 @@
+export * from './usersFollowingAndFollowersService'
+export * from './usersFollowingAndFollowersService.types'
