--- src/services/usersService/index.ts
@@ -1,2 +1,0 @@
-export * from './usersService'
-export * from './usersService.types'
